feat: updating paypal component for v6
