+++ packages/lib/src/components/PayPal/components/PaypalComponent.tsx
@@ -2,9 +2,9 @@ import { h } from 'preact';
 import { useState, useEffect, useCallback } from 'preact/hooks';
 import PaypalButtons from './PaypalButtons';
 import Spinner from '../../internal/Spinner';
-import { getPaypalUrl } from '../utils';
-import { PayPalComponentProps } from '../types';
+import { getPaypalUrl } from '../utils/get-paypal-url';
 import Script from '../../../utils/Script';
+import type { PayPalComponentProps } from '../types';
 
 export default function PaypalComponent({ onApprove, onCancel, onChange, onError, onSubmit, ...props }: PayPalComponentProps) {
     const [status, setStatus] = useState('pending');
