+++ packages/lib/storybook/helpers/checkout-api-calls.ts
@@ -68,7 +68,7 @@ export const patchPaypalOrder = async ({
     pspReference?: string;
     paymentData: string;
     amount: { value: number; currency: string };
-    deliveryMethods: any;
+    deliveryMethods?: any;
 }): Promise<{ paymentData: string }> => {
     if (!(pspReference || sessionId) || !paymentData || !amount.value || !amount.currency) {
         throw Error('PayPal patching order - Field is missing');
