+++ packages/lib/storybook/stories/wallets/PayPal/PayPal.stories.tsx
@@ -1,10 +1,11 @@
 import { Meta, StoryObj } from '@storybook/preact';
 import { PaymentMethodStoryProps } from '../../types';
 import { getStoryContextCheckout } from '../../../utils/get-story-context-checkout';
-import { PayPalElementProps } from '../../../../src/components/PayPal/types';
 import { Container } from '../../Container';
+import Paypal from '../../../../src/components/PayPal';
+import type { PayPalConfiguration } from '../../../../src/components/PayPal/types';
 
-type Story = StoryObj<PaymentMethodStoryProps<PayPalElementProps>>;
+type Story = StoryObj<PaymentMethodStoryProps<PayPalConfiguration>>;
 
 const meta: Meta = {
     title: 'Wallets/Paypal'
@@ -13,13 +14,9 @@ export default meta;
 
 export const Default: Story = {
     render: (args, context) => {
+        const { componentConfiguration } = args;
         const checkout = getStoryContextCheckout(context);
-        return <Container type={'paypal'} componentConfiguration={args.componentConfiguration} checkout={checkout} />;
-    },
-    argTypes: {
-        showPayButton: {
-            control: false,
-            description: 'This property is not supported'
-        }
+        const paypal = new Paypal(checkout, componentConfiguration);
+        return <Container element={paypal} />;
     }
 };
