+++ packages/lib/storybook/stories/wallets/PayPal/PaypalExpressWithReviewPage.stories.tsx
@@ -4,6 +4,7 @@ import { handleSubmit } from '../../../helpers/checkout-handlers';
 import { patchPaypalOrder } from '../../../helpers/checkout-api-calls';
 import { createAdvancedFlowCheckout } from '../../../helpers/create-advanced-checkout';
 import { getDeliveryMethods, getSelectedDeliveryMethodAmount } from './paypal-stories-utils';
+import { PayPal } from '../../../../src';
 
 const meta: Meta = {
     title: 'Wallets/Paypal'
@@ -69,15 +70,14 @@ const Component = () => {
                 showPayButton: true,
                 amount: AMOUNT.value,
                 countryCode: COUNTRY_CODE,
-                shopperLocale: SHOPPER_LOCALE,
-                paymentMethodsConfiguration: {}
+                shopperLocale: SHOPPER_LOCALE
             });
 
             if (!checkout) {
                 return;
             }
 
-            const paypal = checkout.create('paypal', {
+            const paypal = new PayPal(checkout, {
                 isExpress: true,
                 userAction: 'continue',
 
@@ -154,9 +154,11 @@ const Component = () => {
                     component.updatePaymentData(paymentData);
                 },
 
-                onShopperDetails(shopperDetails, paypalOrder, actions) {
-                    console.log('onShopperDetails');
-                    store.setState({ shopperDetails, paypalOrder });
+                onAuthorized(data, actions) {
+                    store.setState({
+                        deliveryAddress: data.deliveryAddress,
+                        paypalOrder: data.authorizedEvent
+                    });
                     actions.resolve();
                 },
