+++ packages/lib/storybook/stories/wallets/PayPal/PaypalExpressWithSessions.stories.tsx
@@ -3,6 +3,7 @@ import { useEffect, useRef } from 'preact/hooks';
 import { patchPaypalOrder } from '../../../helpers/checkout-api-calls';
 import { createSessionsCheckout } from '../../../helpers/create-sessions-checkout';
 import { getDeliveryMethods, getSelectedDeliveryMethodAmount } from './paypal-stories-utils';
+import { PayPal } from '../../../../src';
 
 const meta: Meta = {
     title: 'Wallets/Paypal'
@@ -55,15 +56,14 @@ const Component = () => {
                 showPayButton: true,
                 amount: AMOUNT.value,
                 countryCode: COUNTRY_CODE,
-                shopperLocale: SHOPPER_LOCALE,
-                paymentMethodsConfiguration: {}
+                shopperLocale: SHOPPER_LOCALE
             });
 
             if (!checkout) {
                 return;
             }
 
-            const paypal = checkout.create('paypal', {
+            const paypal = new PayPal(checkout, {
                 isExpress: true,
 
                 blockPayPalVenmoButton: true,
@@ -124,8 +124,8 @@ const Component = () => {
                     component.updatePaymentData(paymentData);
                 },
 
-                onShopperDetails(shopperDetails, paypalOrder, actions) {
-                    console.log('onShopperDetails', shopperDetails, paypalOrder, actions);
+                onAuthorized(data, actions) {
+                    console.log('onShopperDetails', data);
                     actions.resolve();
                 }
             });
